黄色の警告の修正
2024/12/02 11:25
問題が起こったら連絡ください

+++ ComponentFramework/SorceFiles/GameObjects/Component/CameraComponent.cpp
@@ -85,7 +85,7 @@ void CameraComponent::Update()
 	float farPlane = 3.0f;	// ファークリップ
 
 	Matrix projectionMatrix;
-	projectionMatrix = DirectX::XMMatrixOrthographicLH(GameProcess::GetWidth(), GameProcess::GetHeight(), nearPlane, farPlane);
+	projectionMatrix = DirectX::XMMatrixOrthographicLH(static_cast<float>(GameProcess::GetWidth()), static_cast<float>(GameProcess::GetHeight()), nearPlane, farPlane);
